src/load-generator/requirements.txt
Flask==3.1.2
locust-plugins[playwright]==5.0.0
python-json-logger==4.0.0
openfeature-provider-ofrep==0.2.0
